fix: Time component is faded out when user clicks on it many times
If the user clicked on Time Component many times fast,
the component will be faded out. This is because the component is using
`gsap.from` so it starts from zero to the current opacity not to 1.
So, I changed the `gsap.from` to `gsap.fromTo` to fix this issue.

Changes to be committed:
	modified:   apps/toctive-blog/frontend/components/shared/time/time.tsx