fix: station names are hidden on first run
When use make a new trip right after first run, all station in
SeeAllStations screen are hidden because the default language are 'en-US'
not 'en'. So, we split that text and use 'en' part only.

also when the user select any visited station to mark it as visited,
all stations before it should be selected too. I fixed this bug.

Changes to be committed:
	modified:   apps/fast-metro/app/src/components/shared/Station/Station.tsx
	modified:   apps/fast-metro/app/src/screens/SeeAllStations/SeeAllStationsScreen.tsx